requirements.txt
basemap==1.4.1
basemap_data==1.3.2
herbie_data==2024.5.0
matplotlib==3.8.4
numpy==1.26.4
opencv_python==4.9.0.80
opencv_python_headless==4.8.0.74
rasterio==1.3.10
ultralytics==8.2.20
